moved utility types to their own module
